split repo into three parts and add appropriate __init__.py files
