--- covid_moonshot_ml/docking/__init__.py
@@ -1,2 +1,0 @@
-from . import mcs
-from . import docking
