+++ scripts/cdd_to_schema.py
@@ -1,13 +1,9 @@
 import argparse
-import json
 import os
-import pandas
-from rdkit.Chem import MolToSmiles, MolFromSmiles
-import re
 import sys
 
 sys.path.append(f"{os.path.dirname(os.path.abspath(__file__))}/../")
-from covid_moonshot_ml.datasets.cdd import cdd_to_schema, cdd_to_schema_pair
+from asap_dataset_handling.datasets.cdd import cdd_to_schema, cdd_to_schema_pair
 
 
 def get_args():
