+++ scripts/clean_results_csv.py
@@ -4,7 +4,7 @@ import numpy as np
 sys.path.append(
     f"{os.path.dirname(os.path.dirname(os.path.abspath(__file__)))}"
 )
-from covid_moonshot_ml.docking.analysis import DockingResults
+from asap_docking.docking import DockingResults
 
 
 def get_args():
