+++ scripts/docking_data_to_csv.py
@@ -3,7 +3,7 @@ import sys, os, argparse
 sys.path.append(
     f"{os.path.dirname(os.path.dirname(os.path.abspath(__file__)))}"
 )
-from covid_moonshot_ml.docking.analysis import DockingDataset
+from asap_docking.docking import DockingDataset
 
 
 def get_args():
