+++ scripts/run_docking.py
@@ -5,12 +5,12 @@ import pickle as pkl
 import sys
 
 sys.path.append(f"{os.path.dirname(os.path.abspath(__file__))}/../")
-from covid_moonshot_ml.docking.docking import (
+from asap_docking.docking import (
     build_docking_systems,
     parse_xtal,
     run_docking,
 )
-from covid_moonshot_ml.schema import (
+from asap_dataset_handling.schema import (
     ExperimentalCompoundDataUpdate,
     EnantiomerPairList,
 )
