+++ scripts/test-ligand-filter.py
@@ -4,12 +4,12 @@ Function to test implementation of ligand filtering
 
 import sys, os, argparse
 
-import covid_moonshot_ml.datasets.fragalysis
+import asap_dataset_handling.datasets.fragalysis
 
 sys.path.append(
     f"{os.path.dirname(os.path.dirname(os.path.abspath(__file__)))}"
 )
-from covid_moonshot_ml.datasets import utils
+from asap_dataset_handling.datasets import utils
 
 
 def get_args():
@@ -45,8 +45,10 @@ def main():
     args = get_args()
 
     # First, parse the fragalysis directory and
-    sars_xtals = covid_moonshot_ml.datasets.fragalysis.parse_fragalysis_data(
-        args.csv_file, args.fragalysis_dir
+    sars_xtals = (
+        asap_dataset_handling.datasets.fragalysis.parse_fragalysis_data(
+            args.csv_file, args.fragalysis_dir
+        )
     )
 
     # For the compounds for which we have smiles strings, get a dictionary mapping the Compound_ID to the smiles
